fetch data home and shop

+++ lib/models/item.dart
@@ -1,14 +1,14 @@
 class Item {
   String name, image, id;
-  int price, star;
+  double price, star;
   DateTime date;
 
   Item(this.name, this.image, this.star, this.price);
   Item.withId(this.id, this.name, this.image, this.star, this.date);
 
   String get getID => id;
-  int get getPrice => price;
-  int get getStar => star;
+  double get getPrice => price;
+  double get getStar => star;
   String get getName => name;
   String get getImage => image;
   DateTime get getDate => date;
@@ -25,13 +25,13 @@ class Item {
     }
   }
 
-  set setStar(int newStar) {
+  set setStar(double newStar) {
     if (5 >= newStar && newStar >= 0) {
       this.star = newStar;
     }
   }
 
-  set setPrice(int newPrice) {
+  set setPrice(double newPrice) {
     if (newPrice >= 0) {
       this.price = newPrice;
     }
@@ -60,8 +60,8 @@ class Item {
   Item.fromJson(Map<String, dynamic> map) {
     this.id = map['itemID'];
     this.name = map['name'];
-    this.price = int.parse(map['price']);
+    this.price = double.parse(map['price']);
     this.image = map['image'];
-    this.star = int.parse(map['rating']);
+    this.star = double.parse(map['rating']);
   }
 }
